requirements.txt
# Generic
python-dotenv

# Project related
gensim==3.4.0
more-itertools==4.1.0
numpy==1.14.3
scikit-learn==0.19.1
torch==0.4.0
torchtext==0.2.3
torchvision==0.2.1
